refactor: 오타 수정 (#43)
